pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-seclint"
version = "0.1.0"
description = "Static security linter for LLM-powered applications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "xr843" },
]
keywords = ["security", "linting", "llm", "static-analysis", "vulnerability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/xr843/llm-seclint"
Repository = "https://github.com/xr843/llm-seclint"
Issues = "https://github.com/xr843/llm-seclint/issues"

[project.scripts]
llm-seclint = "llm_seclint.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_seclint"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]
